requirements.txt
tts
openai
python-dotenv